added externalRouting
